Task 7.1 Refactoring and New Review-Endpoint
Implements a review endpoint for the CampusCoffee backend.
Reviews are associated with a POS and a user.
They must be approved by 3 users before being published.

Appropriate path mapping and the implemtation of the needed methods is added.
The class ReviewDto is implemented.
The ReviewService interface was added as well as its implementation.
The Logic for the review approval is also implemented.
Commented-out lines have been added back, were appropriate.

Needs to be done: Fix issues